pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "super-agents"
version = "0.1.0"
description = "Multi-sector asset-first agent framework"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "httpx>=0.27",
    "pypdf>=6.0",
    "pydantic>=2.7",
    "PyYAML>=6.0",
    "python-dotenv>=1.0",
    "sqlalchemy>=2.0",
    "pandas>=2.2",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "litellm>=1.40",
    "langfuse>=2.30",
    "streamlit>=1.35",
    "prefect>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "pytest-cov>=5.0",
]

[project.scripts]
super-agents = "super_agents.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
